Fix handling of DISCARD pseudo-register
Previously, the DISCARD register was just lowered to "undef", but could
be `load`ed from and `store`d to. This lead to undefined behaviour (who
would have guessed) and potential crashes.

We now special case `load`s and `store`s from/to DISCARD and crash (i.e.
emit invalid LLVM IR) if the register is used in any other way.