GC: Correctly handle stack offsets relative to base pointer
The previous code (from llvm-statepoint-utils) relied on the "frameSize"
value from LLVM's stackmap. However that value can not be relied on for
correct stack traversal, as the real frame can change during function
execution (e.g. when arguments are passed via the stack, depending on
the generated assembly).

This commit changes the behaviour to always include a frame-pointer,
which can then be used for stack traversal. The used
llvm-statepoint-utils are also adjusted to clearly distinguish between
stackpointer- and basepointer-relative offsets.